% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete.time.markov.R
\name{get.transition.matrix}
\alias{get.transition.matrix}
\title{Computes Transition Probability Matrix}
\usage{
get.transition.matrix(states, option = "prob", output_type = "matrix")
}
\arguments{
\item{states}{a \emph{vector} with values corresponding to states in a Markov chain model.}

\item{option}{if \code{0} or \code{"prob"}, returns transition probabilites; if \code{1} or \code{"count"}, returns counts of each transition from the data.}

\item{output_type}{if \code{"matrix"}, outputs a \emph{matrix}; if \code{"df"}, outputs a \emph{dataframe}.}
}
\value{
A matrix or dataframe containing transition probabilities or counts.
}
\description{
Given a series of states, computes the transition probability matrix or returns a matrix of transition counts.
}
\examples{
# Generate random sequence of states
states <- sample(c('up', 'down', 'same'), 10000, replace=TRUE)
# Create transition probability matrix
x <- get.transition.matrix(states, option='prob', output_type='matrix')
}
