% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete.time.markov.text.R
\name{get.transition.matrix.text}
\alias{get.transition.matrix.text}
\title{Computes Transition Probability Matrix for Text Data}
\usage{
get.transition.matrix.text(
  documents,
  n,
  option = "prob",
  output_type = "matrix",
  punct = "none"
)
}
\arguments{
\item{documents}{a \emph{string} representing a single document or a \emph{vector} of multiple document strings.}

\item{n}{number of words (excluding punctuation) in a state.}

\item{option}{if \code{0} or \code{"prob"}, returns transition probabilities; if \code{1} or \code{"count"}, returns counts of each transition from the data.}

\item{output_type}{if \code{"matrix"}, outputs a \emph{matrix}; if \code{"df"}, outputs a \emph{dataframe}.}

\item{punct}{how the function deals with punctuation. If \code{'none'}, removes all punctuation from input. If \code{'end'}, only punctuation at the end of the document (if any) are considered as states. If \code{'period'}, all punctuation at the end of the document (if any) are considered as periods.}
}
\value{
A matrix or dataframe containing transition probabilities or counts.
}
\description{
Given text data, computes the transition probability matrix between pre-defined states or returns a matrix of transition counts.
}
\details{
Each state is defined as a a group of \code{n} consecutive words. All end-sentence punctuation, depending on input for the \code{punct} parameter, are also considered as states by themselves.

  Assumes each string (each document) contains one sentence each. Also assumes all sentences are in the format "the quick, brown fox jumps over the lazy dog." (i.e. all lowercase, with a period to mark the end of a sentence). Ignores mid-sentence punctuation. If your input text data is not in this format, please clean it accordingly prior to using this function.

  Currently does not support non-English characters (or emojis).

  Use \code{\link{get.steady.state}} on the output of this function to compute steady-state / long-run probabilities for each word/phrase.
}
\seealso{
\code{\link{get.transition.matrix}} for the general, non-text version.
}
