% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete.time.markov.R
\name{get.transitions}
\alias{get.transitions}
\title{Lists All Transitions from Data}
\usage{
get.transitions(states, option = "prob")
}
\arguments{
\item{states}{a \emph{vector} with values corresponding to states in a Markov chain model.}

\item{option}{if \code{0} or \code{"prob"}, returns transition probabilites; if \code{1} or \code{"count"}, returns counts of each transition from the data.}
}
\value{
A dataframe containing either transition probabilities or counts.
}
\description{
Given a series of states, calculates all possible transitions and returns a dataframe of counts or transition probabilities sorted by start state and end state.

  This output is useful for subsequently drawing state transition diagrams using \code{\link{igraph::graph_from_data_frame}}.
}
\examples{
# Generate random sequence of states
states <- sample(c('up', 'down', 'same'), 10000, replace=TRUE)
# Calculate transition probabilities
get.transitions(states, option='prob')

}
\seealso{
\code{\link{get.transition.matrix}}
}
