% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete.time.markov.text.R
\name{get.transitions.text}
\alias{get.transitions.text}
\title{Lists All Transitions from Text Data}
\usage{
get.transitions.text(documents, n, option = "prob", punct = "none")
}
\arguments{
\item{documents}{a \emph{string} representing a single document or a \emph{vector} of multiple document strings.}

\item{n}{number of words (excluding punctuation) in a state.}

\item{option}{if \code{0} or \code{"prob"}, returns transition probabilities; if \code{1} or \code{"count"}, returns counts of each transition from the data.}

\item{punct}{how the function deals with punctuation. If \code{'none'}, removes all punctuation from input. If \code{'end'}, only punctuation at the end of the document (if any) are considered as states. If \code{'period'}, all punctuation at the end of the document (if any) are considered as periods.}
}
\value{
A dataframe containing either transition probabilities or counts.
}
\description{
Given text data, calculates all possible transitions between pre-defined states and returns a dataframe of counts or transition probabilities sorted by start state and end state.
}
\details{
Each state is defined as a a group of \code{n} consecutive words. All end-sentence punctuation, depending on input for the \code{punct} parameter, are also considered as states by themselves.

  Assumes each string (each document) contains one sentence each. Also assumes all sentences are in the format "the quick, brown fox jumps over the lazy dog." (i.e. all lowercase, with a period to mark the end of a sentence). Ignores mid-sentence punctuation. If your input text data is not in this format, please clean it accordingly prior to using this function.

  Currently does not support non-English characters (or emojis).

  The output for this function is useful for subsequently drawing state transition diagrams using \code{\link{igraph::graph_from_data_frame}}.

  To get a transition probability matrix instead, use \code{\link{get.transition.matrix.text}}.
}
\seealso{
\code{\link{get.transitions}} for the general, non-text version.
}
